refactor: 551292 - FieldsetQuestion abstract class
